+++ components/selfupdater/source/googleplay/src/main/kotlin/com/flipperdevices/selfupdater/googleplay/api/SelfUpdaterGooglePlayApi.kt
@@ -1,9 +1,10 @@
 package com.flipperdevices.selfupdater.googleplay.api
 
-import android.app.Activity
 import android.content.Context
+import com.flipperdevices.core.activityholder.CurrentActivityHolder
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.error
 import com.flipperdevices.core.log.info
 import com.flipperdevices.inappnotification.api.InAppNotificationStorage
 import com.flipperdevices.inappnotification.api.model.InAppNotification
@@ -49,12 +50,18 @@ class SelfUpdaterGooglePlayApi @Inject constructor(
         info { "Process checkout new update" }
         appUpdateManager.registerListener(updateListener)
         appUpdateInfoTask.addOnSuccessListener { appUpdateInfo ->
+            val currentActivity = CurrentActivityHolder.getCurrentActivity()
+            if (currentActivity == null) {
+                error { "Current activity is null, skip update check" }
+                return@addOnSuccessListener
+            }
+
             if (isUpdateAvailable(appUpdateInfo)) {
                 info { "New update available, suggest to update" }
                 appUpdateManager.startUpdateFlowForResult(
                     appUpdateInfo,
                     AppUpdateType.FLEXIBLE,
-                    context as Activity,
+                    currentActivity,
                     UPDATE_CODE
                 )
             }