+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - [FIX] Fix open key screen in global scope
 - [FIX] Wait for disconnect while update
 - [FIX] Fix mfkey32
+- [FIX] Use CurrentActivityHolder for get activity in updater(replace Context cast)
 - [REFACTOR] Migrate bottom bar to compose navigation
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues